more refactors to /shared
