+++ examples/application/landDrainageConsent.json
@@ -41,7 +41,9 @@
               "last": "of Mortain"
             },
             "email": "countbobby@email.org",
-            "phone": "012345678901"
+            "phone": {
+              "primary": "012345678901"
+            }
           }
         },
         {
@@ -58,7 +60,9 @@
               "last": "De Leybourne"
             },
             "email": "baronsarerevolting@email.org",
-            "phone": "098765432109",
+            "phone": {
+              "primary": "098765432109"
+            },
             "company": {
               "name": "The Barons' Revolt"
             }
