+++ types/schemas/preApplication/data/Applicant.ts
@@ -1,9 +1,6 @@
-import {Email} from '../../../shared/utils';
-import {
-  Agent,
-  SiteContact,
-  UserAddress,
-} from '../../application/data/Applicant';
+import {UserAddress} from '../../../shared/Addresses';
+import {ContactDetails} from '../../../shared/Contacts';
+import {Agent, SiteContact} from '../../application/data/Applicant';
 
 /**
  * @id #Applicant
@@ -15,20 +12,8 @@ export type Applicant = BaseApplicant | Agent;
  * @id #BaseApplicant
  * @description Information about the user who completed the application for themself, or information about the person who the user applied on behalf of
  */
-export interface BaseApplicant {
+export type BaseApplicant = ContactDetails & {
   type: 'individual' | 'company' | 'charity' | 'public' | 'parishCouncil';
-  name: {
-    title?: string;
-    first: string;
-    last: string;
-  };
-  email: Email;
-  phone: {
-    primary: string;
-  };
-  company?: {
-    name: string;
-  };
   address: UserAddress;
   siteContact: SiteContact;
-}
+};
